added trainer test module and docs
